requirements.txt
google-cloud-translate
python-dotenv
altair==5.0.1
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
blis==0.7.10
cachetools==5.3.1
catalogue==2.0.8
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.2.0
click==8.1.6
confection==0.1.0
cymem==2.0.7
de-core-news-md @ https://github.com/explosion/spacy-models/releases/download/de_core_news_md-3.6.0/de_core_news_md-3.6.0-py3-none-any.whl#sha256=c289fe1bfcc3e3670f2e0e849e8b8c83ecb83cc70045522a1e06a56b291ed30d
de-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/de_core_news_sm-3.6.0/de_core_news_sm-3.6.0-py3-none-any.whl#sha256=cdded59f89eacf348c30a79f9a2485ab21e154ef300cc02c7ebe87dea149c457
decorator==5.1.1
Flask==2.3.2
gitdb==4.0.10
GitPython==3.1.32
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-core==2.3.3
google-cloud-translate==3.11.2
googleapis-common-protos==1.59.1
googletrans==4.0.0rc1
grpcio==1.56.2
grpcio-status==1.56.2
gunicorn==21.2.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
langcodes==3.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
murmurhash==1.0.9
nltk==3.8.1
numpy==1.25.1
packaging==23.1
pandas==2.0.3
pathy==0.10.2
Pillow==9.5.0
preshed==3.0.8
proto-plus==1.22.3
protobuf==4.23.4
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.11
pydeck==0.8.0
Pygments==2.15.1
Pympler==1.0.1
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
referencing==0.30.0
regex==2023.6.3
requests==2.31.0
rfc3986==1.5.0
rich==13.4.2
rpds-py==0.9.2
rsa==4.9
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
spacy==3.6.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.7
streamlit==1.25.0
tenacity==8.2.2
thinc==8.1.10
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
typer==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==1.26.16
validators==0.20.0
wasabi==1.1.2
Werkzeug==2.3.6
zipp==3.16.2
